+++ e2e/addon/mock/testdata/example/definitions/helm.yaml
