+++ e2e/addon/mock/testdata/example/metadata.yaml
