+++ e2e/addon/mock/testdata/example/resources/configmap.cue
