+++ e2e/addon/mock/testdata/example/resources/parameter.cue
