+++ e2e/addon/mock/testdata/example/resources/service/source-controller.yaml
