+++ e2e/addon/mock/testdata/example/template.yaml
