+++ e2e/addon/mock/testdata/fluxcd/definitions/helm-release-def.yaml
