+++ e2e/addon/mock/testdata/fluxcd/definitions/json-patch.yaml
