+++ e2e/addon/mock/testdata/fluxcd/definitions/kustomize.yaml
