+++ e2e/addon/mock/testdata/fluxcd/definitions/strategy-merge.yaml
