+++ e2e/addon/mock/testdata/fluxcd/metadata.yaml
