+++ e2e/addon/mock/testdata/fluxcd/resources/crds/bucket.yaml
