+++ e2e/addon/mock/testdata/fluxcd/resources/crds/git-repo.yaml
