+++ e2e/addon/mock/testdata/fluxcd/resources/crds/helm-release.yaml
