+++ e2e/addon/mock/testdata/fluxcd/resources/crds/helm-repo.yaml
