+++ e2e/addon/mock/testdata/fluxcd/resources/crds/kustomize.yaml
