+++ e2e/addon/mock/testdata/fluxcd/resources/deployment/helm-controller.yaml
