+++ e2e/addon/mock/testdata/fluxcd/resources/deployment/kustomize-controller.yaml
