+++ e2e/addon/mock/testdata/fluxcd/resources/deployment/source-controller.yaml
