+++ e2e/addon/mock/testdata/fluxcd/resources/rbac/binding-cluster-admin.yaml
