+++ e2e/addon/mock/testdata/fluxcd/resources/rbac/binding-crd-controller.yaml
