+++ e2e/addon/mock/testdata/fluxcd/resources/rbac/crd-controller.yaml
