+++ e2e/addon/mock/testdata/fluxcd/resources/rbac/rbac-helm-controller.yaml
