+++ e2e/addon/mock/testdata/fluxcd/resources/rbac/rbac-kustomize-controller.yaml
