+++ e2e/addon/mock/testdata/fluxcd/resources/rbac/rbac-source-controller.yaml
