+++ e2e/addon/mock/testdata/fluxcd/resources/service/svc-source-controller.yaml
