+++ e2e/addon/mock/testdata/fluxcd/resources/service/webhook-receiver.yaml
