+++ e2e/addon/mock/testdata/fluxcd/schemas/component-uischema-helm.yaml
