+++ e2e/addon/mock/testdata/fluxcd/template.yaml
