+++ e2e/addon/mock/testdata/terraform-alibaba/definitions/terraform-alibaba-ack.yaml
