+++ e2e/addon/mock/testdata/terraform-alibaba/metadata.yaml
