+++ e2e/addon/mock/testdata/terraform-alibaba/resources/alibaba-account-creds.cue
