+++ e2e/addon/mock/testdata/terraform-alibaba/resources/alibaba-provider.cue
