+++ e2e/addon/mock/testdata/terraform-alibaba/resources/parameter.cue
