+++ e2e/addon/mock/testdata/terraform-alibaba/template.yaml
