+++ e2e/addon/mock/testdata/terraform/metadata.yaml
