+++ e2e/addon/mock/testdata/test-addon/metadata.yaml
