+++ e2e/addon/mock/testdata/test-addon/template.yaml
